Move FollowButton to general components
