+++ frontend/src/components/general/FollowButton.js
@@ -1,7 +1,7 @@
 import { Button, CircularProgress, Link, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import React from "react";
-import ButtonIcon from "./ButtonIcon";
+import ButtonIcon from "../project/Buttons/ButtonIcon";
 
 const useStyles = makeStyles((theme) => ({
   followButtonContainer: {
